correctly handle constant accesses to banked memories
