--- file-tests/should-lower/if-one-branch.fuse
@@ -1,5 +1,0 @@
-let y: bit<32>;
-let x: bit<32>;
-
-if (x == 0) { x := 2; }
-y := x;
