+++ src/main/scala/passes/LowerUnroll.scala
@@ -1,7 +1,7 @@
 package fuselang.passes
 
 import scala.{PartialFunction => PF}
-import fuselang.Utils.RichOption
+import fuselang.Utils.{RichOption, cartesianProduct}
 import fuselang.common._
 import Syntax.{OpConstructor => OC}
 import Transformer._
@@ -19,32 +19,18 @@ object LowerUnroll extends PartialTransformer {
     prefix + curIdx
   }
 
-  private def cartesianProduct[T](llst: Seq[Seq[T]]): Seq[Seq[T]] = {
-    def pel(e: T, ll: Seq[Seq[T]], a: Seq[Seq[T]] = Nil): Seq[Seq[T]] =
-      ll match {
-        case Nil => a.reverse
-        case x +: xs => pel(e, xs, (e +: x) +: a)
-      }
-
-    llst match {
-      case Nil => Nil
-      case x +: Nil => x.map(Seq(_))
-      case x +: _ =>
-        x match {
-          case Nil => Nil
-          case _ =>
-            llst
-              .foldRight(Seq(x))((l, a) => l.flatMap(x => pel(x, a)))
-              .map(_.dropRight(x.size))
-        }
-    }
-  }
-
   private def genViewAccessExpr(suffix: Suffix, idx: Expr): Expr =
     suffix match {
       case Aligned(factor, e2) => (EInt(factor) * e2) + idx
       case Rotation(e) => e + idx
     }
+
+  // Key for transformers: A sequence that tracks the index expression
+  // and the bank implied by that expression.
+  private type TKey = Seq[(Expr, Option[Int])]
+  // Value generated by transformer: Mapping from the banks in each dimension
+  // to the expression that access array.
+  private type TVal = Map[Seq[Int], Expr]
 
   /**
     * A function for transforming an access expression where each index
@@ -64,31 +50,49 @@ object LowerUnroll extends PartialTransformer {
     * The function returns a map from a Seq[Int] (bank numbers) to the
     * access expression that corresponds to it.
     */
-  private type TKey = Seq[(Expr, Option[Int])]
-  private type TVal = Map[Seq[Int], Expr]
   case class ViewTransformer(val t: TKey => TVal) extends AnyVal
   object ViewTransformer {
+
+    /**
+      * Define a transformer for a base memory.
+      */
     def fromArray(id: Id, ta: TArray) = {
-      // Get the name of all the memories
       val t = (idxs: Seq[(Expr, Option[Int])]) => {
+        // If any of the indices are constants, scale them to the index in the
+        // bank.
+        // Morally, accesses like A[2] (where A: float[4 bank 2]) implicitly
+        // use a view over the base physical access: A{0}[1].
+        val normIdxs = idxs
+          .zip(ta.dims)
+          .map({
+            case ((idx, bank), (_, totalBanks)) =>
+              idx match {
+                case EInt(i, base) => (EInt(i / totalBanks, base), bank)
+                case _ => (idx, bank)
+              }
+          })
+
+        // For each dimension, get the bank implied by the access expression.
+        // If the expression doesn't imply a bank, it touches all banks.
         val allBanks: Seq[Seq[Int]] = ta.dims
-          .zip(idxs)
+          .zip(normIdxs)
           .map({
             case ((_, arrBank), (_, bank)) => {
+              // Either get the bank implied by this expression or
+              // get enumerate every bank in this dimension.
               bank.map(Seq(_)).getOrElse(0 until arrBank)
             }
           })
 
+        // Enumerate all the bank access expressions implied by this
+        // access.
         cartesianProduct(allBanks)
           .map(banks => {
             (
               banks,
               EArrAccess(
                 Id(id.v + banks.mkString("_")),
-                idxs
-                  .map(_._1)
-                  .zip(ta.dims.map(_._2))
-                  .map({ case (i, b) => { println(s"$i, $b"); (i / EInt(b, 10)) }})
+                normIdxs.map(_._1)
               )
             )
           })
@@ -97,6 +101,9 @@ object LowerUnroll extends PartialTransformer {
       ViewTransformer(t)
     }
 
+    /**
+      * Define a transformer for view built on top of a memory.
+      */
     def fromView(dims: Seq[DimSpec], v: CView): ViewTransformer = {
       val t = (idxs: Seq[(Expr, Option[Int])]) => {
         if (idxs.length != dims.length) {
@@ -164,7 +171,7 @@ object LowerUnroll extends PartialTransformer {
       this
     }
 
-    override def withScopeAndRet[V](inScope: Env =>  (V, Env)) = {
+    override def withScopeAndRet[V](inScope: Env => (V, Env)) = {
       val (ret, nEnv) = inScope(this.copy(rewrites = rewrites.addScope))
       val rws = nEnv.rewrites.endScope match {
         case Some((_, rs)) => rs
@@ -570,6 +577,8 @@ object LowerUnroll extends PartialTransformer {
               arrDims,
               (e) => c.copy(lhs = e)
             )(env)
+          // Recursively rewrite command generated for this physical
+          // access.
           rewriteC(newCmd)(nEnv)
         }
         case _ => throw Impossible("Not an LHS")
